chore(ag): extend user group invitation with expires_at property

See merge request camac-ng/camac-ng!11174

+++ ember-ebau-core/addon/controllers/service-permissions/invitations.js
@@ -4,6 +4,7 @@ import { service } from "@ember/service";
 import { tracked } from "@glimmer/tracking";
 import { task, timeout } from "ember-concurrency";
 import { query } from "ember-data-resources";
+import { confirm } from "ember-uikit";
 
 import paginatedQuery from "ember-ebau-core/resources/paginated";
 
@@ -11,6 +12,9 @@ export default class ServicePermissionsInvitationsController extends Controller 
   @service store;
   @service session;
   @service ebauModules;
+  @service intl;
+  @service fetch;
+  @service notification;
 
   @tracked search = "";
   @tracked inGroup = null;
@@ -35,7 +39,41 @@ export default class ServicePermissionsInvitationsController extends Controller 
   delete = task({ drop: true }, async (userGroupInvitation, event) => {
     event.preventDefault();
 
-    await userGroupInvitation.destroyRecord();
+    if (
+      await confirm(
+        this.intl.t("service-permissions.invitation-delete-confirm"),
+        {
+          i18n: {
+            ok: this.intl.t("global.delete"),
+          },
+        },
+      )
+    ) {
+      await userGroupInvitation.destroyRecord();
+    }
+  });
+
+  renew = task({ drop: true }, async (userGroupInvitation, event) => {
+    event.preventDefault();
+
+    try {
+      await this.fetch.fetch(
+        `/api/v1/user-group-invitations/${userGroupInvitation.id}/renew`,
+        {
+          method: "POST",
+        },
+      );
+
+      await userGroupInvitation.reload();
+
+      this.notification.success(
+        this.intl.t("service-permissions.invitation-renew-success"),
+      );
+    } catch {
+      this.notification.danger(
+        this.intl.t("service-permissions.invitation-renew-error"),
+      );
+    }
   });
 
   updateSearch = task({ restartable: true }, async (event) => {
